@@ -1,8 +1,11 @@
 module github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/eventhub/armeventhub
 
-go 1.16
+go 1.18
+
+require github.com/Azure/azure-sdk-for-go/sdk/azcore v0.23.0
 
 require (
-	github.com/Azure/azure-sdk-for-go/sdk/azcore v0.21.0
-	github.com/Azure/azure-sdk-for-go/sdk/azidentity v0.13.0
+	github.com/Azure/azure-sdk-for-go/sdk/internal v0.9.1 // indirect
+	golang.org/x/net v0.0.0-20210610132358-84b48f89b13b // indirect
+	golang.org/x/text v0.3.6 // indirect
 )
